feat(refactor): move pageData to data/ folder
